cleaned up code a bit
