--- .nyc_output/2d047fbb-2e14-49be-86c3-8d7961855f19.json
@@ -1,1 +1,0 @@
-{}
