--- .nyc_output/4e638c82-24ef-414e-ac31-398aa3d3cb7f.json
@@ -1,1 +1,0 @@
-{}
