--- .nyc_output/875d4526-5808-446d-a675-6cf8725ba198.json
@@ -1,1 +1,0 @@
-{}
