--- .nyc_output/ff89306f-80b2-4458-b300-553a613b1c1a.json
@@ -1,1 +1,0 @@
-{}
