--- .nyc_output/processinfo/2d047fbb-2e14-49be-86c3-8d7961855f19.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":23116,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-prefix.js"],"execArgv":[],"cwd":"C:\\Repos\\vscode-mssql","time":1741240541541,"ppid":8576,"coverageFilename":"C:\\Repos\\vscode-mssql\\.nyc_output\\2d047fbb-2e14-49be-86c3-8d7961855f19.json","externalId":"","uuid":"2d047fbb-2e14-49be-86c3-8d7961855f19","files":[]}
