--- .nyc_output/processinfo/4e638c82-24ef-414e-ac31-398aa3d3cb7f.json
@@ -1,1 +1,0 @@
-{"parent":"875d4526-5808-446d-a675-6cf8725ba198","pid":29688,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Repos\\vscode-mssql\\node_modules\\@playwright\\test\\cli.js","test"],"execArgv":[],"cwd":"C:\\Repos\\vscode-mssql","time":1741240543316,"ppid":25016,"coverageFilename":"C:\\Repos\\vscode-mssql\\.nyc_output\\4e638c82-24ef-414e-ac31-398aa3d3cb7f.json","externalId":"","uuid":"4e638c82-24ef-414e-ac31-398aa3d3cb7f","files":[]}
