--- .nyc_output/processinfo/875d4526-5808-446d-a675-6cf8725ba198.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":7220,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\nvm\\v20.18.0\\node_modules\\npm\\bin\\npm-cli.js","exec","--","playwright","test"],"execArgv":[],"cwd":"C:\\Repos\\vscode-mssql","time":1741240541969,"ppid":19684,"coverageFilename":"C:\\Repos\\vscode-mssql\\.nyc_output\\875d4526-5808-446d-a675-6cf8725ba198.json","externalId":"","uuid":"875d4526-5808-446d-a675-6cf8725ba198","files":[]}
