--- .nyc_output/processinfo/ff89306f-80b2-4458-b300-553a613b1c1a.json
@@ -1,1 +1,0 @@
-{"parent":"4e638c82-24ef-414e-ac31-398aa3d3cb7f","pid":32704,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Repos\\vscode-mssql\\node_modules\\playwright\\lib\\common\\process.js"],"execArgv":[],"cwd":"C:\\Repos\\vscode-mssql","time":1741240545192,"ppid":29688,"coverageFilename":"C:\\Repos\\vscode-mssql\\.nyc_output\\ff89306f-80b2-4458-b300-553a613b1c1a.json","externalId":"","uuid":"ff89306f-80b2-4458-b300-553a613b1c1a","files":[]}
