+++ test/e2e/baseFixtures.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+// taken from mxschimtt: https://github.com/mxschmitt/playwright-test-coverage/blob/main/e2e/baseFixtures.ts
+
 import * as fs from "fs";
 import * as path from "path";
 import * as crypto from "crypto";
